added floating point number parser along with testing framework
A floating point number parser was a good place to start to get working
a testing framework with coverage testing.  Details about this project
were add to the README file.

An initial main CMakeLists.txt file was added, which checks the CMake
version, sets the C++17 standard, turns on all warnings and errors,
verifies that building is not occurring in the source directory (a build
sub-directory is acceptable), checks the build type and echos it
(allows an environment variable to be set), enables testing, looks for
the code coverage tools (lcov and genhtml) for a Debug build, adds the
add coverage test function (adds a test and coverage test), and adds the
Catch (unit test platform) and Parser sub-directories.

The Catch sub-directory contains the single Catch header file along with
a catch_main.cpp source file (prevents long compiling of the Catch
header file for main() function generation as it only needs to be
compiled once) and it's license file.  The CATCH_SOURCES cmake variable
is set in the parent scope so this can be used for the unit tests.

The Parser sub-directory contains the source files for parsing floating
point numbers along with unit tests and code coverage.  There is also a
Parse error class for holding an error message along with a column where
the error was detected.