added simple identifier parser
To support the PRINT and END commands, will need to parse for an
identifier.  Since these commands are only letters, identifiers of all
alphabetic characters are parsed.  This parser will be extended as
needed going forward to parse other kinds of identifiers (for example,
variables contains digits after the first letter for variable names,
underscore characters for variable names, a trailing '(' for functions,
a trailing dollar sign for string variables, etc.).