simplified test coverage script to speed it up a bit
The desired coverage data can be extracted directly when it is first
generated, so the additional steps to extract data for the directory,
and then extracting again for each file and appending it together are
not necessary.  The speeds up the coverage test a bit (about 25%), but
is still quite slow, probably due to the newer coverage data format with
new GCC versions.