added program operation instruction code class
A Thunder program will consist of an array of program words.  A program
word will contain an instruction operation code (opcode) or an operand
for an opcode.  There will be an ocode for each unique program element
(command, operator, function, constant, variable, etc.).  To create
unique opcode valuess for each element, there will be an instance of the
OpCode class each being assigned a unique opcode value when instanced.