added program word class
A Thunder program will consist of an array of program words.  A program
word will contain an instruction code or an operand value.

The code coverage was pulling in too many files; the OpCode source files
were also being included in the Word coverage results.  There was a
problem in the test coverage script where it was including all source
files, not just the files being tested.  The script was modified to only
include the files passed.  The name of the coverage output file was also
changed since this name appears in the coverage HTML output as the test
name.