added compiling and recreating of simple commands
The new ProgramCode class is for holding code, which is a simple vector
of program words.  For now that is a single line of code, though later
it could hold an entire program (with a higher level class managing the
lines within the code).  Its constructor will compile a line from an
input stream.

In order to convert command keywords into a code, a static map was added
to the CommandOpCode class from opcode value to keyword along with a
static getKeyword() function.

In order to test the compiling code without adding access function to
the words array for testing only, a recreateLine() function was added to
the ProgramCode class.  This function takes an offset into the words
array.  For now zero is used for this offset since the words array only
contains a single line, but eventually the words array will contain many
lines of code, so an offset is needed to indicate which line.