added compile function to command opcode instances
These functions will be used to compile a particular command.  The
compile functions (pointers) are stored in a map of opcode value to
compile function pointer.

corrected implementation of static command opcode info

There was a problem with the implementation of the static codes and
keywords maps, which was discovered when the print and end opcode
instances were moved from the unit tests to the code source file (i.e.
made static instances).  The issue was the opcode instance constructors
were called before the map constructors were called to initialize the
maps, and it crashed upon trying to store into the uninitialized maps.

This was resolved by creating a private Command class containing these
maps along with the command() function that returns the single instance
of this class.  This way, the Command (and the maps) are initialized on
demand (in other words, before they are used by a command opcode
constructor).

A static compile() function was added to CommandOpCode to compile a
command.  The opcode (used to lookup a compile function) and a reference
to the ProgramCode instance is passed to the compile function.  The
addOpCode() function was added to ProgramCode so that the compile
functions can add their code's opcode to the program.

Removed an unnecessary test from the ProgramWord unit test for a command
opcode.