started expanding recreator functionality
Added a new Recreator class, which will contain functions for the opcode
recreate functions.  For now this is just a function to add the command
keyword, and a function to get the recreated line.  Added a Table class
for holding the recreate function pointers for the opcodes.  The is a
static recreate function to the OpCode class.