started expending compiler functionality
Added new Compiler class, which will contain functions for the opcode
compile functions providing access to the program code.  For now this
includes functions for compiling a line (main call), compiling an
expression (which does nothing yet) and adding an opcode to the program
code.  It contains a reference to the program code and in instances of
the Parser class for parsing.  Some of the compiling code was moved from
the ProgramCode class.  The opcode compile functions now take a
reference to a Compiler instance instead of the ProgramCode.