added support for expressions with a single constant
For now the compile expression function skips white space, and parses
a number.  If a number was parsed, a constant number opcode is added to
the program code, the value of the constant is added to the program
data and the index of the constant is added to the program code.

This required new parser skip white space functionality.  Recreation of
constant numbers was added.  The recreator functionality was expanded
where it iterates program code until one of the opcodes (a command)
marks that the recreation [of the line] is done.  The recreator add
command keyword was extended to prepend the keyword if there is already
characters on the line along with a space separator.