added runner support to execute program code
Added Runner class to run a program.  This class will take the program
code instance and an output stream where program output will go.  All of
the current run functions were put into the runner source file (for
now). Added support for run functions to the OpCodes class (all opcodes
will have run functions).

Changed ProgramCode constructor taking an input stream to compile a
line to the compile line function so that it can be called without
constructing a new instance.

Changed compile expression to return a boolean to whether a valid
expression (for now number constant) was compiled.

Changed compile print function to check if an expression was compile and
if yes, add a print number opcode to the program.  And added support for
the print number opcode including a recreate nothing function (for now).