added very simple console class
The Console class will be the interface between user and the program.
The code structure has been implemented that issues the ready prompt
along with a colon indicating a command can be entered.  Only the quit
command has been implemented, which causes the console to terminate.
The Console class takes a generic input and output stream.  This allows
if to be fed with test input and for the output to be captured.

The console tests uses input and output stream streams.  A thunder
executable was also created, which just instantiates a Console with the
standard console input and output streams, and runs the console.