added simple list command support to console
This is just for the automatic 'end' statement in a program.  This
required a new ProgramUnit class, which will contain the program code
along with other program information (like where each line will begin
in the program, not yet implement).  It contains a recreate line
function for creating a line for a give line number (line numbers are
arbitrary in a Thunder program and always start at one and count
sequentially be ones).  A line number is added to the line with a space
(line numbers are base one externally, but base zero internally).  There
is also a private function for obtaining an offset for a line (only
returns zero at the moment since a program will only have the single
automatic 'end' statement).

Besides adding a ProgramUnit instance to Console, a list command ('l')
was also added to the Console class, which currently just prints a
single (first and only) line.  The 'Ready' prompt was changed to only
be issued at the beginning.  If will also be issued after certain
commands, but not the list command.