changed how coverage is tested and reduced the number of unit tests
As the number of unit tests grew, the time to run coverage on these
tests grew.  There was also a lot of duplication of tests among the
unit tests.

It was therefore decided to run coverage on all of the unit tests at one
(in other words, reset the coverage data, run all of the unit tests,
then measure coverage of all of the code for these unit tests).  This
allowed the removal of many of the [duplicate] unit tests, leaving two
for the parser directory (identifier and number), and one each for the
compiler, program (runner), and console directories.

This reduced testing time from over 3 minutes to about a minute.  Build
time is also reduced due to less unit tests.  A side effect of these
changes is that the unit tests need to be serially (but only for the
Debug build required to run coverage), but this is not an issue since
the unit tests very quickly (less than a tenth of a second combined).