replaced catch2 with version 3
This change was made in the effort to reduce coverage testing time,
which was growing for each new unit test due to the amount of coverage
data produced by the Catch header file (that is ultimately filtered
out).  Version 3 of Catch moves away from an all-inclusive single header
file to smaller header files and a library.  This also means that Catch
was removed from this project and needs to be installed on the system
(install instructions are on the wiki).

The Catch directory was removed.  A find package for Catch2 version 3
was added to the main CMake script with the required option.  For each
unit test, the Catch2 with main library was added.  In each unit test
source file, the include statement was updated to the main Catch2 header
file along with an include of the exception matchers header file to use
the REQUIRE_THROWS_WITH macro.

This reduced coverage testing time from around a minute to mere three
seconds.