added insert command to console
Support for the 'i' insert command was added to Console.  Multiple lines
can be inserted (terminated with a blank line, but for now each
additional insert just inserts at the beginning of the program.  The
'l' list command was extended to list all the lines in the program
instead of just the first line.

After an entered line is input and compiled, it needs to be inserted
into the program, which requires adding the resources (just constant
numbers at the moment) from the program line to the program and updating
the indexes in the operands.  The OperandTranslator class handles this
adding of resources and 'translating' the operand indexes of the program
line before these words are inserted into the program.  To support
multiple lines in the ProgramUnit, new ProgramLine class contains the
starting offset for each line.