changed recreator mechanism for determining and of line
Previously an is done flag was used to stop the recreate line loop,
which was set when a command keyword is added to the line.  Now a line
view is passed that in additional to containing the offset of the line,
also contains the size of the line.

This was done for two reasons.  First to allow a part of code to be
created (for example just an expression, which will be useful during
the develop of expressions), and second to allow the possibily for
multiple statements per line (a feature that may or may not be added in
the future, possibly also as a limited feature allow in certain
circumstances).

The compiler unit tests needed to be updated to pass a ProgramView to
the recreate line function with the correct size.

Also refactored the ProgramLines::View embedded structure in a separate
ProgramView structure in its own header (some code the requires
ProgramView does not also required ProgramLines, so it made sense that
it doesn't belong inside ProgramLines).