added compiling of unary operators and unary expressions
A unary expression is defined as an expression the contains zero or more
unary operators followed by a single operand.  Only the '-' negate unary
operator has been implemented so far.  A new Operators class was
implement to hold the information for operator opcodes.  Unary operators
will only be single characters.

A new expression unit test program was created.  The ExpressionTester
class was created for testing single expressions, which includes member
functions for compiling, recreating and running an expression.  The
compile function takes a string of the expression to compile, the
recreate function recreates the expression and returns a string, and the
run function runes the functions and returns the result of the
expression.  The run function first adds an END opcode so that execution
will stop immediately after the expression code, and returns the top of
the stack.