added binary operator support (add only)
Added binary operators to the Operators class.  For now binary operators
are single characters (but this will need to change singe some binary
operators are two characters).  Changed the single line string in the
Recreate class to a string stack.  Changed the single number in the
Runner class to a number stack.