added proper handling of operator precedences
Added a Precedence enumeration class.  Added a precedences map from an
opcode value.  Defined precedences to the negate and add operators
(operator initialization lists now contain a precedence value that is
used to initialize the precedences map).  Added a static function to
get a precedence for an opcode.

In the Compiler class, changed there OpCode stack to an Operator stack
where the Operator class contains the OpCode and Precedence.  Changed
the flush operator (formally opcode) stack to take an argument
indicating higher or same precedence operator that should be flushed
from the stack (lower precedence operators are not flushed).  Added a
Lowest Precedence value that is used at the end of an expression to
flush all remaining operators.  Added a Bottom Precedence value (lower
than Lowest) along with a bottom opcode that is put on the bottom of
the operator stack (to eliminate the need to check for an empty stack).


Refactored the Runner class handling of the run stack where access to
the top of the stack is directly accessible (replaced the get and set
top number functions with a top number reference function).  Moved the
stack access functions to the header so they will be inlined providing
better performance for running.