added support for parentheses in expressions
Parentheses are used to alter the order of executing operators from
their normal precedences.

Opening parentheses located in expressions where unary operators are
found so is considered a unary operator.  Its precedence is just above
the bottom so that no other operators will flush it from the operator
stack.  In addition to pushing it on the operator stack, an entry is
also pushed on a new sub-expression stack.

Closing parentheses are location in expression where binary operators
are found so is considered a binary operator.  Its precedence is just
above that of an open parentheses so that all operators are flushed from
the operator stack except the open parenthesis.  Instead of adding a
closing parenthesis on the operator stack, the open parenthesis is
popped from the stack, the sub-expression stack is popped and it loops
back to look for another binary operator.

To stop parsing before a closing parenthesis without an open parentheses
(so that the caller can report the appropriate error), the new
sub-expression stack is used.  For now this stack doesn't actually have
any contents.  A new other argument was added to the get binary opcode
routine that uses this argument to determine if the operator is allowed.
For now, only the closing parenthesis operator is checked.  If the
sub-expression is empty, a closing parenthesis operator is not allowed.

Corrected a problem with getting the column where -1 was being returning
when getting the position after the last character of the stream was
read.  This was fixed by getting the current column in the Parser peek
next character before peeking at the next character.  This is stored in
a new member of the Parser class, which is return when requested.  This
change required column member to be removed from the NumberParser, which
was essentially doing the same thing.

Changed the Recreator class string stack to contain both a string and
precedence value.  This is used doing binary operator recreation to
determine whether parentheses need to be put around a sub-expression
when the precedence is lower (left-side) or lower or same (right-side).

Removed not currently used OpCode == and != overload functions.
Corrected bug in print number run routine where a constant buffer wasn't
large enough for the largest possible float point number.