simplified determination of other operator in sub-expressions
Even though the determination of the other operator for sub-expression
wasn't that involved currently (check if this stack is empty), this
logic would have grown more complex with the addition of functions,
arrays, strings, etc.

However, at the time an entry is pushed to the sub-expression stack, it
will be known what the other operator will be, so this will be pushed
to the sub-expression (which was just empty entries).  Now the top
entry of this stack will have the other operator value, which can be
passed directly to the get binary opcode routine.  A None entry is
pushed on the stack at the beginning.

(It may be a concern in the future that these bottom entries are not
being removed from the stack, which upon repeated calls with the same
compiler instance will cause these stacks to grow.)