added most of the math functions
Support for most of the basic math functions including ABS, INT, SGN,
SQR, LOG, EXP, COS, SIN, TAN and ATN, but not RND.

A new Token class was added for holding an identifier with attributes
(for now just whether it has a parenthesis).  A new parse token function
was added to the parser for getting an identifier followed by an
optional parenthesis.

A new compile operand function was added to the Compiler class, which
for now gets a token, gets the function opcode (the only type of operand
at the moment).  If the token has a parenthesis, the function opcode is
pushed to the operator stack with a function precedence (low to avoid
getting flushed until the closing parenthesis) and the closing
parenthesis other operator type is pushed to the sub-expression stack.
If the token does not have a parenthesis, the function is pushed to the
operator stack with a unary function precedence (high so that other
operators will flush it).  In other words the function will act like a
unary operator and no closing parenthesis will be expected (only will
allow a single operand for the argument of the function.

A new recreate function routine was added for handling all function
recreations.  The argument of the function will always be surrounded by
parentheses following the function name regardless of whether
parentheses were entered.  (Entry without parentheses is only a
convenience).