added random function
The RND is the final of the basic math functions.  The RND function has
two forms, one with an argument and one without.  Without an argument,
RND returns a random number from 0 to 1.  With an argument, the random
number is scaled (multiplied) by the argument.  For special case
argument of 0, RND will return the previous random number returned
(including if it was scaled).  Because of the no argument form, the RND
function with arguments requires parentheses to be used (so the compiler
know which form is being used).

A new alternate opcode mechanism was added to the Functions class.  Only
RND currently has an alternate opcode.  A number of arguments mechanism
was also added to Function (where most functions have 1 argument except
the no argument RND function).  The repetitive 'function' names on
member functions, arguments and members were removed.

In Compiler class compile operand routine, if the [function] token has
no parenthesis, if now gets the alternate code for the function, and if
is has one, the code is immediately added to the program (assumed to
have no arguments) and new 'Added' status is returned to indicate an
operand has been added to tell the compile unary expression routine it
can return (like it does with number constants).  Otherwise as before,
it stacks the function opcode and returns 'Sub-expression' status but
does not push on to the sub-expression stack so it won't expect a
closing parenthesis.

The recreate function routine was updated to first get the number of
arguments the function has, and if none, just pushes the name of the
function onto the string stack, otherwise it behaves as before (adding
parentheses around the argument).

A new local RandomNumberGenerator class was implemented in the runner
source file, which contains a standard random number and will produce
uniformly distributed random numbers taking scale factor to multiple
by.  It also keeps the last random number draw.  An instance of this
class is created each time the program is run so that the same random
sequence is produced.  Because this instance is held in a unique
pointer, a default Runner destructor needed to be defined in the source
file.

Two new functions were added to the Runner class, one for getting a
random draw, and one to set the top of the run stack to the last random
draw.