corrected and improved error reporting and added a few command options
Line numbers can now be entered before the command in addition to after
the command (for now just with the insert command).  A line number only
is now interpreted as an insert at that line number.  A line number with
a statement will insert the statement at that line number and then
enters insert mode after that line.

Statements with errors will now report the error instead of crashing
back to the shell.  Errors messages now contain a visual pointer to the
location of the error on the input line.

A new CommandParser class was implemented to contain routines for
parsing command lines in the Console, with function for get a line
number, a command (single letter), getting the remainder of the line,
checking if at end of line and records columns and lengths for error
reporting.

The ParseError class was renamed to just Error as it will apply to more
than just parsing errors (probably also need to move it from the Parser
directory).  The column member type was changed from long to std::size_t
to eliminate need for casting (from std::streampos).

Implemented a convenience container find function that returns
optionally the value found (no value if not found), so that comparison
to the end of container is not needed.
the